+++ modules/core/src/main/java/net/ashald/envfile/EnvFileParser.java
@@ -10,5 +10,4 @@ public interface EnvFileParser {
 
     @NotNull Map<String, String> process(@NotNull String path, @NotNull Map<String, String> source) throws EnvFileErrorException, IOException;
 
-    @NotNull String getTitle();
 }