+++ modules/core/src/main/java/net/ashald/envfile/EnvVarsProviderFactory.java
@@ -1,0 +1,12 @@
+package net.ashald.envfile;
+
+import org.jetbrains.annotations.NotNull;
+
+
+public interface EnvVarsProviderFactory {
+
+    @NotNull EnvFileParser createParser();
+
+    @NotNull String getTitle();
+
+}