+++ modules/core/src/main/java/net/ashald/envfile/providers/dotenv/DotEnvFileParserFactory.java
@@ -1,0 +1,19 @@
+package net.ashald.envfile.providers.dotenv;
+
+import net.ashald.envfile.EnvFileParser;
+import net.ashald.envfile.EnvVarsProviderFactory;
+import org.jetbrains.annotations.NotNull;
+
+public class DotEnvFileParserFactory implements EnvVarsProviderFactory {
+
+    @Override
+    public @NotNull EnvFileParser createParser() {
+        return new DotEnvFileParser();
+    }
+
+    @Override
+    public @NotNull String getTitle() {
+        return ".env";
+    }
+
+}