+++ modules/core/src/main/java/net/ashald/envfile/providers/yaml/YamlEnvFileParserFactory.java
@@ -1,0 +1,19 @@
+package net.ashald.envfile.providers.yaml;
+
+import net.ashald.envfile.EnvFileParser;
+import net.ashald.envfile.EnvVarsProviderFactory;
+import org.jetbrains.annotations.NotNull;
+
+public class YamlEnvFileParserFactory implements EnvVarsProviderFactory {
+
+    @Override
+    public @NotNull EnvFileParser createParser() {
+        return new YamlEnvFileParser();
+    }
+
+    @NotNull
+    public String getTitle() {
+        return "JSON/YAML";
+    }
+
+}