+++ modules/core/src/test/java/net/ashald/envfile/providers/dotenv/DotEnvFileParserTest.java
@@ -1,4 +1,4 @@
-package net.ashald.envfile.parsers;
+package net.ashald.envfile.providers.dotenv;
 import net.ashald.envfile.EnvFileErrorException;
 import org.junit.Assert;
 import org.junit.Test;